requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
distro==1.6.0
Django==3.2.8
django-mathfilters==1.0.0
h11==0.12.0
icalendar==4.0.9
idna==3.3
numpy==1.21.3
outcome==1.1.0
pandas==1.3.4
pycparser==2.20
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
tabula-py==2.3.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
xlrd==2.0.1
